requirements.txt
pyserial>=3.5
PyQt5>=5.15.0
pyqtgraph>=0.13.0
pandas>=1.5.0
numpy>=1.23.0
